Color conv attempt 2. Start consolidating simple color conversion functions in a single header+cpp.
