+++ Common/Common.vcxproj.filters
@@ -45,6 +45,7 @@
     <ClInclude Include="MathUtil.h" />
     <ClInclude Include="BitSet.h" />
     <ClInclude Include="CodeBlock.h" />
+    <ClInclude Include="ColorConv.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="stdafx.cpp" />
@@ -79,6 +80,7 @@
     <ClCompile Include="ArmThunk.cpp" />
     <ClCompile Include="MipsEmitter.cpp" />
     <ClCompile Include="Arm64Emitter.cpp" />
+    <ClCompile Include="ColorConv.cpp" />
   </ItemGroup>
   <ItemGroup>
     <None Include="CMakeLists.txt" />
