+++ GPU/Directx9/TextureScalerDX9.cpp
@@ -26,6 +26,7 @@
 #include "GPU/Directx9/TextureScalerDX9.h"
 
 #include "Core/Config.h"
+#include "Common/ColorConv.h"
 #include "Common/Common.h"
 #include "Common/Log.h"
 #include "Common/MsgHandler.h"
